Reorg endpoints into /user/info, and /user/oauth/google
