+++ multinet/__init__.py
@@ -9,7 +9,8 @@ from sentry_sdk.integrations.flask import FlaskIntegration
 
 from typing import Optional, MutableMapping, Any, Tuple, Union
 
-import multinet.auth.google as google
+from multinet import auth
+from multinet.auth import google
 from multinet import api
 from multinet.db import register_legacy_workspaces
 from multinet import uploaders, downloaders
@@ -41,9 +42,10 @@ def create_app(config: Optional[MutableMapping] = None) -> Flask:
     app.register_blueprint(downloaders.csv.bp, url_prefix="/api")
     app.register_blueprint(downloaders.d3_json.bp, url_prefix="/api")
 
-    app.register_blueprint(google.bp, url_prefix="/google")
-    google.init_oauth(app)
+    app.register_blueprint(auth.bp, url_prefix="/user")
+    app.register_blueprint(google.bp, url_prefix="/user/oauth/google")
 
+    google.init_oauth(app)
     register_legacy_workspaces()
 
     # Register error handler.
